requirements.txt
pandas==1.1.0
streamlit==0.64.0
plotly==4.9.0
Pillow==7.2.0
